Extract the edit component
